Don't do interceptions at the very end of traveling between planets.
Right after arriving at the destination planet, the code would do one more
check for a possible interception.  The chance of this happening is small, but
if it happened, it would leave the game in an inconsistent state.  This might
have caused the game to immediately start a real mission, which would loop
indefinitely.