Completely eliminate the chance of grinding at interceptions.
It was previously just ineffective. Now it's impossible.

The reason I'm doing this is when you *can* grind (in this case, by
constantly getting free plasma and selling it), it kind of feels
like an obligation to do so.

It's still possible to grind within missions whose progress depends
on your pace, which aren't timed and keep generating more enemies
until you win. Namely: the missile boat mission, the miner mission,
and the third boss. (There might be others.) However, grinding in
these situations is quite dangerous and likely to take away the much
larger shield bonus at the end, and plus if you mess up when trying
to do this you're set back quite a bit, so I think these are sufficiently
worthless activities that no one will do them. (I never did.)