Added a new original Project: Starfighter theme: Android Oppression
This is a new theme for Starfighter! This version represents mystery
and the ominous presence of WEAPCO. Other versions will be used for
other situations, and a sort of combination between this and Last
Cyber Dance will serve as the music for the final Kline battle. That's
the theory anyway.