Adjusted instrument balance of Android's Last Dance, added some percussion.
The quiet part at the end was nearly inaudible in the game, making
it sound like silence even though it wasn't. Insertion of a little
light percussion in that section should avoid this. Also adjusted
audio levels of the instruments so that they're better balanced, and
actually committing the changed mmpz file (I forgot to do so last
time).