Corrected a conflation of characters with code points.
Fixes #36, and also fixes a problem where spaces would get inserted
to the start of subsequent lines by line breaks via the same mechanism.
The problem was the string was being indexed (by code-point) using
an index value for characters, which worked fine for English as only
ASCII is used in that case, but not other languages.